Moved from a special language setting to a general select setting
