--- src/renderer/settings/setting/language-setting.js
@@ -1,30 +1,0 @@
-
-import React from 'react';
-import { Select, MenuItem } from '@material-ui/core';
-
-import { supported_languages } from '../../../common/local/locale';
-
-const styles = {
-    select: {
-        width: '-webkit-fill-available',
-        textAlign: 'left',
-        height: '3.5rem',
-    },
-};
-
-function LanguageSetting(props) {
-    const onChange = (e) => {
-        props.onUpdate(e.target.value);
-    };
-    return (
-        <div>
-            <Select style={styles.select} value={props.option} variant="outlined" onChange={onChange}>
-                {Object.keys(supported_languages).map((id) => (
-                    <MenuItem key={id} value={id}>{supported_languages[id]}</MenuItem>
-                ))}
-            </Select>
-        </div>
-    );
-}
-
-export default LanguageSetting;
