+++ src/renderer/settings/setting/setting-types.js
@@ -1,7 +1,7 @@
 
 import BooleanSetting from './boolean-setting';
 import ColorSetting from './color-setting';
-import LanguageSetting from './language-setting';
+import SelectSetting from './select-setting';
 import ShortcutSetting from './shortcut-setting';
 import SizeSetting from './size-setting';
 import QualitySetting from './quality-setting';
@@ -13,7 +13,7 @@ import CodeSetting from './code-setting';
 const setting_types = {
     boolean: BooleanSetting,
     color: ColorSetting,
-    language: LanguageSetting,
+    select: SelectSetting,
     shortcut: ShortcutSetting,
     size: SizeSetting,
     quality: QualitySetting,
